+++ CHANGELOG.md
@@ -19,5 +19,6 @@ and this project adheres to
 - ✨(OIDC Provider) Add a `guest` claim and the `account` scope
 - 🔒(authentication) make the logout view POST-only
 - 👽(authentication) handle Social Auth's login views requiring POST requests
+- 🦖(OIDC Provider) Introspection endpoint can fall back to PSA backends
 
 [unreleased]: https://github.com/suitenumerique/docs/compare/main